parallel karatsuba working

+++ numArray.h
@@ -1,6 +1,6 @@
 #ifndef NUMARRAY_H_   /* Include guard */
 #define NUMARRAY_H_
 
-char *file_to_array(long length, char *file);
+char *file_to_array(long long length, char *file);
 
 #endif // NUMARRAY_H_
